+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## 1.0.1 (15th June 2015)
+* Improved the ability to enable the variable product column in the licence table.
+* Removed commented out code that is not used.
+* Removed closing PHP tags at the end of the files.
+* Removed superglobal $GLOBALS directly in the main plugin file.
+* Updated the text-domain name.
+* Updated the name of the plugin class.
+* Updated the README.md and readme.txt files.
+
 ## 1.0.0 (15th April 2015)
 
 * Initial Release